requirements.txt
dj-database-url==0.4.2
django-cors-headers==3.10.1
django-environ==0.8.0
django-filter==2.4.0
djangorestframework==3.13.0
djangorestframework-xml==2.0.0
drf-nested-routers==0.93.4 
gunicorn==20.0.4
psycopg2==2.9.3
psycopg2-binary==2.9.2
Django==4.0.4
pytest ==7.1.2
pytest-django == 4.5.2
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
drf-yasg == 1.21.3
pillow == 9.2.0